ESP32_Fingerprint_Enroll
This is the configuration using adafruits fingerprint library to enroll employees via the ESP32. 
To do from here: 
Connect to  the node.js application via httpGet to pull the data for creating an employee. 
Allow esp32 to talk to mongoDB to store employees. 
Once the image is taken and database document is created you will see this result on the web application. 
From here the details will be added to the employee. 
Following on i must write a program to constantly search for fingerprints to allow active monitoring.  
Using time stamps i will create another database document to store this data to be used in the pay calculation side of the application.