Employee error
Added controller to do a log, also added file I/O that caused an error on the show employees. Database connection working and tested. Ran a populate script. Tested add employee in create and this too is working. So that leave 2 possible causes. 1) Error in the find method in 2) Error in the angular to display employees.